Major mods
Removed the src/adafruit_portalbase folder with the modified wifi_esp32s2.py script. We don't need any longer to use a modified adafruit module. So we are now complete in line with the modules maintained by Adafruit. Freeing us from possible future update problems.
For this we modified code.py on various points.
Checked the result. It is running as intended.